requirements.txt
yt-dlp>=2024.3.10
opencv-python>=4.9.0
ultralytics>=8.1.0
pytesseract>=0.3.10
Pillow>=10.2.0
easyocr>=1.7.0
numpy<2.0.0
instaloader>=4.10.0
requests>=2.31.0
beautifulsoup4>=4.12.3
playwright>=1.42.0
ffmpeg-python>=0.2.0
rich>=13.7.0
tqdm>=4.66.2
python-dotenv>=1.0.1
pytest>=7.4.0
